setup.sh: do not override custom files

+++ setup.sh
@@ -1,9 +1,15 @@
 #!/usr/bin/env bash
 
+set -euo pipefail
+IFS=$'\n\t'
+
 dir="$(realpath $(dirname "$0"))"
 
 cd "$dir"
 for f in $(ls -A home); do
+	if [[ ! -L "$HOME/$f" ]]; then
+		continue
+	fi
 	unlink "$HOME/$f"
 	ln -s "$dir/home/$f" "$HOME/$f"
 done
